# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/auth/login": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login and get token.",
                "parameters": [
                    {
                        "description": "request format",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RegisterUserCommand"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.RegisterUserResponse"
                        }
                    }
                }
            }
        },
        "/v1/auth/register": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register new user.",
                "parameters": [
                    {
                        "description": "request format",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RegisterUserCommand"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.RegisterUserResponse"
                        }
                    }
                }
            }
        },
        "/video/api/v1/comment/admin/delete": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment"
                ],
                "summary": "Force delete comment by admin role",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/comment.ForceDeleteCommentCommand"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/comment.ForceDeleteCommentResponse"
                        }
                    }
                }
            }
        },
        "/video/api/v1/comment/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment"
                ],
                "summary": "Create comment.",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/comment.CreateCommentCommand"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/comment.CreateCommentResponse"
                        }
                    }
                }
            }
        },
        "/video/api/v1/comment/delete": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment"
                ],
                "summary": "Delete comment by normal user",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/comment.DeleteCommentCommand"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/comment.DeleteCommentResponse"
                        }
                    }
                }
            }
        },
        "/video/api/v1/comment/find": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment"
                ],
                "summary": "Find comments by video id.",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "videoId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/comment.FindByVideoResponse"
                        }
                    }
                }
            }
        },
        "/video/api/v1/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "video"
                ],
                "summary": "Create video post. This API is called after the video upload. Should bring the uploaded video uuid in the request.",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/video.CreateVideoCommand"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/video.CreateVideoResponse"
                        }
                    }
                }
            }
        },
        "/video/api/v1/video/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "video"
                ],
                "summary": "Upload video.",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/video.UploadVideoCommand"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/video.UploadVideoResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.RegisterUserCommand": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "example@gmail.com"
                },
                "password": {
                    "type": "string",
                    "example": "password"
                },
                "username": {
                    "type": "string",
                    "example": "username01"
                }
            }
        },
        "auth.RegisterUserResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "comment.CreateCommentCommand": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "videoId": {
                    "type": "integer"
                }
            }
        },
        "comment.CreateCommentResponse": {
            "type": "object",
            "properties": {
                "commend_id": {
                    "type": "string"
                }
            }
        },
        "comment.DeleteCommentCommand": {
            "type": "object",
            "properties": {
                "commentId": {
                    "type": "string"
                }
            }
        },
        "comment.DeleteCommentResponse": {
            "type": "object"
        },
        "comment.FindByVideoResponse": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.VideoComment"
                    }
                }
            }
        },
        "comment.ForceDeleteCommentCommand": {
            "type": "object",
            "properties": {
                "commentId": {
                    "type": "string"
                }
            }
        },
        "comment.ForceDeleteCommentResponse": {
            "type": "object",
            "properties": {
                "deleteCount": {
                    "type": "integer"
                }
            }
        },
        "entity.VideoComment": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "integer"
                },
                "comment": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "videoId": {
                    "type": "integer"
                }
            }
        },
        "video.CreateVideoCommand": {
            "type": "object",
            "properties": {
                "authorId": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "video.CreateVideoResponse": {
            "type": "object"
        },
        "video.UploadVideoCommand": {
            "type": "object",
            "properties": {
                "uploaderId": {
                    "type": "integer"
                }
            }
        },
        "video.UploadVideoResponse": {
            "type": "object",
            "properties": {
                "uuid": {
                    "type": "string"
                }
            }
        }
    }
}